코드리뷰 반영 -> 모듈화 재진행
