+++ javascript/constants.js
@@ -1,0 +1,21 @@
+// email input
+const inputEmail = document.querySelector(".sign-input");
+const inputBoxEmail = document.querySelector(".sign-input-box");
+// password input
+const inputPassword = document.querySelector(".sign-input#password");
+const inputBoxPassword = document.querySelector(".sign-input-box.sign-password");
+// login button
+const loginButton = document.querySelector(".cta");
+// eye button & img
+const eyeButton = document.querySelector(".eye-button");
+const eyeImg = document.querySelector("#eye-id");
+
+// password-check input
+const inputPasswordCheck = document.querySelector(".sign-input#password-check");
+const inputBoxPasswordCheck = document.querySelector(".sign-input-box.sign-password-check"); 
+// eye button & img (비밀번호 확인)
+const eyeButtonCheck = document.querySelector(".eye-button-check");
+const eyeImgCheck = document.querySelector("#eye-id-check");
+
+export {inputEmail, inputBoxEmail, inputPassword, inputBoxPassword, loginButton, eyeButton, eyeImg,
+        inputPasswordCheck, inputBoxPasswordCheck, eyeButtonCheck, eyeImgCheck};
