+++ javascript/regex.js
@@ -1,0 +1,6 @@
+const USER_EMAIL = "test@codeit.com";
+const USER_PASSWORD = "codeit101";
+const REGEX_EMAIL = /^[A-Za-z0-9_\.\-]+@[A-Za-z0-9\-]+\.[A-Za-z0-9\-]+/;
+const REGEX_PASSWORD = /^(?=.*[a-zA-Z])(?=.*[0-9]).{8,}$/;
+
+export {USER_EMAIL, USER_PASSWORD, REGEX_EMAIL, REGEX_PASSWORD}
