+++ javascript/signup.js
@@ -1,61 +1,53 @@
-import * as sign from './signin.js';
+import {USER_EMAIL} from './regex.js';
+import {isValidEmail, isValidPassword, showErrorMessage, removeMessage, clickEyeIcon, eventClickEye} from './utils.js'
+import {inputEmail, inputBoxEmail, inputPassword, inputBoxPassword, loginButton, eyeButton,
+        inputPasswordCheck, inputBoxPasswordCheck, eyeButtonCheck, eyeImgCheck} from './constants.js';
 
-// password-check input
-const inputPasswordCheck = document.querySelector(".sign-input#password-check");
-const inputBoxPasswordCheck = document.querySelector(".sign-input-box.sign-password-check"); 
-// eye button & img (비밀번호 확인)
-const eyeButtonCheck = document.querySelector(".eye-button-check");
-const eyeImgCheck = document.querySelector("#eye-id-check");
-
-function isValidPassword(passwordValue) {
-    const expPassword = /^(?=.*[a-zA-Z])(?=.*[0-9]).{8,}$/;
-    return expPassword.test(passwordValue);    
-}
 // focus-out email
 function eventFocusOutEmail() {
     // no email value
-    if(sign.inputEmail.value === "") {
-        sign.showErrorMessage(sign.inputBoxEmail, "이메일을 입력해주세요");
+    if(inputEmail.value === "") {
+        showErrorMessage(inputBoxEmail, "이메일을 입력해주세요");
     }
     // not email form
-    else if(!sign.isValidEmail(sign.inputEmail.value)) {
-        sign.showErrorMessage(sign.inputBoxEmail, "올바른 이메일 주소가 아닙니다.");
+    else if(!isValidEmail(inputEmail.value)) {
+        showErrorMessage(inputBoxEmail, "올바른 이메일 주소가 아닙니다.");
     }
-    else if(sign.inputEmail.value === sign.USER_EMAIL) {
-        sign.showErrorMessage(sign.inputBoxEmail, "이미 사용 중인 이메일입니다.");
+    else if(inputEmail.value === USER_EMAIL) {
+        showErrorMessage(inputBoxEmail, "이미 사용 중인 이메일입니다.");
     }
     // error message remove
-    else if(sign.isValidEmail(sign.inputEmail.value)) {
-        sign.removeMessage(sign.inputBoxEmail);
+    else if(isValidEmail(inputEmail.value)) {
+        removeMessage(inputBoxEmail);
     }
 }
 // focus-out password
 function eventFocusOutPassword() {
     // no password value
-    if(sign.inputPassword.value === "") {
-        sign.showErrorMessage(sign.inputBoxPassword, "비밀번호를 입력해주세요");
+    if(inputPassword.value === "") {
+        showErrorMessage(inputBoxPassword, "비밀번호를 입력해주세요");
     }
-    else if(!isValidPassword(sign.inputPassword.value)) {
-        sign.showErrorMessage(sign.inputBoxPassword, "비밀번호는 영문, 숫자 조합 8자 이상 입력해주세요.");
+    else if(!isValidPassword(inputPassword.value)) {
+        showErrorMessage(inputBoxPassword, "비밀번호는 영문, 숫자 조합 8자 이상 입력해주세요.");
     }
     else {
-        sign.removeMessage(sign.inputBoxPassword);
+        removeMessage(inputBoxPassword);
     }
 }
 function eventFocusOutPasswordCheck() {
     // no password value
-    if(inputPasswordCheck.value !== sign.inputPassword.value) {
-        sign.showErrorMessage(inputBoxPasswordCheck, "비밀번호가 일치하지 않아요.");
+    if(inputPasswordCheck.value !== inputPassword.value) {
+        showErrorMessage(inputBoxPasswordCheck, "비밀번호가 일치하지 않아요.");
     }
     else {
-        sign.removeMessage(inputBoxPasswordCheck);
+        removeMessage(inputBoxPasswordCheck);
     }
 }
 // signup button click
 function eventClickBtnSignup() {
-    if( sign.isValidEmail(sign.inputEmail.value)
-        && isValidPassword(sign.inputPassword.value)
-        && inputPasswordCheck.value === sign.inputPassword.value ) {
+    if( isValidEmail(inputEmail.value)
+        && isValidPassword(inputPassword.value)
+        && inputPasswordCheck.value === inputPassword.value ) {
             window.location.href = "/folder";
         }
     else {
@@ -70,13 +62,13 @@ function eventKeyUpEnterSignup(e) {
     }
 }
 function eventClickEyeCheck() {
-    sign.clickEyeIcon(inputPasswordCheck, eyeImgCheck);
+    clickEyeIcon(inputPasswordCheck, eyeImgCheck);
 }
 
-sign.inputEmail.addEventListener('focusout', eventFocusOutEmail);
-sign.inputPassword.addEventListener('focusout', eventFocusOutPassword);
+inputEmail.addEventListener('focusout', eventFocusOutEmail);
+inputPassword.addEventListener('focusout', eventFocusOutPassword);
 inputPasswordCheck.addEventListener('focusout', eventFocusOutPasswordCheck);
-sign.loginButton.addEventListener('click', eventClickBtnSignup);
+loginButton.addEventListener('click', eventClickBtnSignup);
 document.addEventListener('keyup', eventKeyUpEnterSignup);
-sign.eyeButton.addEventListener('click', sign.eventClickEye);
+eyeButton.addEventListener('click', eventClickEye);
 eyeButtonCheck.addEventListener('click', eventClickEyeCheck);
